refactor(core-rs): split per-run locking out of advance.rs (300-line rule)
RunLocks (in_flight serialization + cancel watch channels) lifts verbatim
into run_locks.rs; advance.rs drops 343 -> 290 lines. No behavior change —
the full concurrency suite (cancel/resume/single-flight) stays green.
