+++ packages/core-rs/crates/mainframe-automations/src/engine/advance.rs
@@ -3,10 +3,7 @@
 //! reconcile, retry); concurrent calls for one run serialize on a per-run
 //! lock so a step never executes twice from a race.
 
-use std::collections::HashMap;
-use std::sync::{Arc, Mutex as StdMutex, MutexGuard};
-
-use tokio::sync::{Mutex as TokioMutex, watch};
+use std::sync::Arc;
 
 use crate::domain::{AutomationDefinition, RunActionStep, Step, find_step_by_id};
 use crate::error::StoreError;
@@ -14,6 +11,7 @@ use crate::ports::{AutomationEvent, Clock, EventSink, to_run_summary};
 use crate::store::{RunRecord, RunStore, RunTriggerContext, StepStatus, TerminalStatus};
 
 use super::checkpoint::fail_step_entry;
+use super::run_locks::{RunLocks, cancel_requested};
 use super::walk::{WalkCtx, walk_steps};
 use super::{BoxFuture, RunAdvancer, RunFinalizedHook, VerbPorts, WalkResult};
 
@@ -44,16 +42,14 @@ pub struct InterpreterDeps {
 
 pub struct Interpreter {
     pub(crate) deps: InterpreterDeps,
-    in_flight: StdMutex<HashMap<String, Arc<TokioMutex<()>>>>,
-    cancels: StdMutex<HashMap<String, watch::Sender<bool>>>,
+    locks: RunLocks,
 }
 
 impl Interpreter {
     pub fn new(deps: InterpreterDeps) -> Self {
         Self {
             deps,
-            in_flight: StdMutex::new(HashMap::new()),
-            cancels: StdMutex::new(HashMap::new()),
+            locks: RunLocks::new(),
         }
     }
 
@@ -75,12 +71,12 @@ impl Interpreter {
 
     /// Serialized per-run entry point.
     pub async fn advance(&self, run_id: &str) -> Result<(), StoreError> {
-        let lock = self.lease(run_id);
+        let lock = self.locks.lease(run_id);
         let result = {
             let _guard = lock.lock().await;
             self.advance_inner(run_id).await
         };
-        self.release(run_id, lock);
+        self.locks.release(run_id, lock);
         result
     }
 
@@ -90,9 +86,7 @@ impl Interpreter {
     /// registrations so a chat that finishes later finds nothing to wake.
     /// A run that is already terminal is a silent no-op.
     pub async fn cancel_run(&self, run_id: &str) -> Result<(), StoreError> {
-        if let Some(tx) = lock_map(&self.cancels).get(run_id) {
-            let _ = tx.send(true);
-        }
+        self.locks.request_cancel(run_id);
         match self
             .deps
             .store
@@ -128,7 +122,7 @@ impl Interpreter {
             return Ok(());
         };
 
-        let mut cancel_rx = self.register_cancel(run_id);
+        let mut cancel_rx = self.locks.register_cancel(run_id);
         let walk_outcome = {
             let ctx = WalkCtx {
                 run_id: &run.id,
@@ -148,7 +142,7 @@ impl Interpreter {
                 _ = cancel_requested(&mut cancel_rx) => None,
             }
         };
-        lock_map(&self.cancels).remove(run_id);
+        self.locks.clear_cancel(run_id);
 
         let Some(walk_result) = walk_outcome else {
             // Aborted: cancel_run owns the finalize.
@@ -268,31 +262,6 @@ impl Interpreter {
         }
         Ok(())
     }
-
-    fn lease(&self, run_id: &str) -> Arc<TokioMutex<()>> {
-        lock_map(&self.in_flight)
-            .entry(run_id.to_string())
-            .or_default()
-            .clone()
-    }
-
-    /// Drops the per-run lock entry once nobody else holds it (checked under
-    /// the map mutex, so no new clone can race the removal).
-    fn release(&self, run_id: &str, lock: Arc<TokioMutex<()>>) {
-        let mut map = lock_map(&self.in_flight);
-        if map
-            .get(run_id)
-            .is_some_and(|entry| Arc::ptr_eq(entry, &lock) && Arc::strong_count(entry) == 2)
-        {
-            map.remove(run_id);
-        }
-    }
-
-    fn register_cancel(&self, run_id: &str) -> watch::Receiver<bool> {
-        let (tx, rx) = watch::channel(false);
-        lock_map(&self.cancels).insert(run_id.to_string(), tx);
-        rx
-    }
 }
 
 impl RunAdvancer for Interpreter {
@@ -310,28 +279,6 @@ impl RunAdvancer for Interpreter {
                 .await
         })
     }
-}
-
-/// Resolves when cancel is requested; pends forever otherwise (the walk
-/// branch of the `select!` finishes first).
-async fn cancel_requested(rx: &mut watch::Receiver<bool>) {
-    if *rx.borrow() {
-        return;
-    }
-    while rx.changed().await.is_ok() {
-        if *rx.borrow() {
-            return;
-        }
-    }
-    std::future::pending::<()>().await
-}
-
-/// A poisoned map mutex only means another task panicked mid-insert; the
-/// map itself is still coherent (db.rs precedent).
-fn lock_map<T>(mutex: &StdMutex<T>) -> MutexGuard<'_, T> {
-    mutex
-        .lock()
-        .unwrap_or_else(|poisoned| poisoned.into_inner())
 }
 
 // PORT STATUS: greenfield (docs/plans/2026-07-12-automations-v2-rust-engine.md T4.1, A8), not a TS port
