+++ packages/core-rs/crates/mainframe-automations/src/engine/mod.rs
@@ -12,6 +12,7 @@ mod deadline;
 pub(crate) mod expects;
 pub mod notify_verb;
 pub mod run_action_verb;
+mod run_locks;
 pub(crate) mod walk;
 
 pub use advance::{AgentWaitRegistry, Interpreter, InterpreterDeps};
